Ampliacion del back,entities y repositories. ajuste a uml v13. 2 querys Categorias.1 Query Empresa. 3 Query Sucursal. Borrado Imagen y Usuario

+++ Clases_bidireccionales/src/main/java/com/example/buensaborback/domain/entities/ArticuloManufacturadoDetalle.java
@@ -14,7 +14,7 @@ import lombok.*;
 @ToString
 @Builder
 public class ArticuloManufacturadoDetalle extends Base {
-    private Double cantidad;
+    private Integer cantidad;
     
     @ManyToOne(cascade = CascadeType.PERSIST)
     @JoinColumn(name = "articulo_insumo_id")
